requirements.txt
langchain
openai
pandas
sqlalchemy
python-dotenv
kaggle
serpapi
requests
tqdm
typeguard
langchain-community 
langchain-experimental
langchain-groq
pyprojroot









